+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version latest (0.5.0?, 2019-09-24?)
 
 - ...
 - enh: display an instance's tags in the optimization table
+- feat: add `--filter-tags` option
 
 
 Version 0.5.0 (2019-09-25)